add(providers): add provider for structuring user's cloud file system

+++ mobile/src/services/LocalStorage.ts
@@ -2,26 +2,25 @@ import AsyncLocalStorage from "@react-native-async-storage/async-storage";
 
 class LocalStorage {
     static async getItem(key: string) {
-        const value =
-            (await AsyncLocalStorage.getItem(key)) || !!localStorage
-                ? localStorage.getItem(key)
-                : null;
+        const value = localStorage
+            ? localStorage.getItem(key)
+            : await AsyncLocalStorage.getItem(key);
         return value;
     }
 
     static async setItem(key: string, value: string) {
         await AsyncLocalStorage.setItem(key, value);
-        localStorage.setItem(key, value);
+        if (localStorage) localStorage.setItem(key, value);
     }
 
     static async removeItem(key: string) {
         await AsyncLocalStorage.removeItem(key);
-        localStorage.removeItem(key);
+        if (localStorage) localStorage.removeItem(key);
     }
 
     static async clear() {
         await AsyncLocalStorage.clear();
-        localStorage.clear();
+        if (localStorage) localStorage.clear();
     }
 }
